Navigated to end of maze but does not recognize the finish
Created several objects to identify the direction needed to move from one place to another and also to identify the coordinates of a target cell based on a given movement and starting location. Added code to return to the previous decision point if a loop is identified or if a dead end is reached. Used a stack to store the alternative paths that were not taken. Mouse is able to navigate until the finish but does not recognize that it has reached the end; instead it retraces its movements back to the start of the maze.